cert: lift DistributionPointName bits to x509.
Previously we implemented support for parsing the optional
`DistributionPointName` from a certificate's CRL distribution points
extension with code/types in `cert.rs`.

While the top-level extension differs slightly for the CRL issuing
distribution point extension, it shares the same encoding for the
`distributionPoint` field holding a `DistributionPointName.`

In anticipation of sharing this code between `cert.rs` and `crl.rs` this
commit lifts the existing code from `cert.rs` into `x509.rs` where it
can more easily be shared between the two, similar to other bits common
to both certs and CRLs.
